Multi-threaded entity splitter
Implemented buffering for OSM element (OsmBuffer class) that allows us
to implement reading and writing in separate threads.

Using this to provide a threaded implementation of the EntitySplitter.
It reads data in one thread, puts read entities into a OsmBuffer and
writes them to disk in another thread which reads from the OsmBuffer.

The threaded version can be activated on the CLI using the '-threaded'
option.